refactor: move magic to magicApi
